+++ packages/optimizely-sdk/lib/core/project_config/project_config_manager.tests.js
@@ -16,11 +16,12 @@
 import sinon from 'sinon';
 import { assert } from 'chai';
 import { cloneDeep } from 'lodash';
+
 import { sprintf } from '@optimizely/js-sdk-utils';
 import * as logging from '@optimizely/js-sdk-logging';
-import * as datafileManager from '../datafile_manager';
+import * as datafileManager from '@optimizely/js-sdk-datafile-manager';
 
-import projectConfig from './index';
+import * as projectConfig from './index';
 import { ERROR_MESSAGES, LOG_MESSAGES } from '../../utils/enums';
 import testData from '../../tests/test_data';
 import * as projectConfigManager from './project_config_manager';
@@ -31,7 +32,7 @@ describe('lib/core/project_config/project_config_manager', function() {
   var globalStubErrorHandler;
   var stubLogHandler;
   beforeEach(function() {
-    sinon.stub(datafileManager, 'createHttpPollingDatafileManager').returns({
+    sinon.stub(datafileManager, 'HttpPollingDatafileManager').returns({
       start: sinon.stub(),
       stop: sinon.stub(),
       get: sinon.stub().returns(null),
@@ -50,7 +51,7 @@ describe('lib/core/project_config/project_config_manager', function() {
   });
 
   afterEach(function() {
-    datafileManager.createHttpPollingDatafileManager.restore();
+    datafileManager.HttpPollingDatafileManager.restore();
     logging.resetErrorHandler();
     logging.resetLogger();
   });
@@ -185,9 +186,9 @@ describe('lib/core/project_config/project_config_manager', function() {
           updateInterval: 10000,
         },
       });
-      sinon.assert.calledOnce(datafileManager.createHttpPollingDatafileManager);
+      sinon.assert.calledOnce(datafileManager.HttpPollingDatafileManager);
       sinon.assert.calledWithExactly(
-        datafileManager.createHttpPollingDatafileManager,
+        datafileManager.HttpPollingDatafileManager,
         sinon.match({
           datafile: JSON.stringify(config),
           sdkKey: '12345',
@@ -200,7 +201,7 @@ describe('lib/core/project_config/project_config_manager', function() {
     describe('when constructed with sdkKey and without datafile', function() {
       it('updates itself when the datafile manager is ready, fulfills its onReady promise with a successful result, and then emits updates', function() {
         var configWithFeatures = testData.getTestProjectConfigWithFeatures();
-        datafileManager.createHttpPollingDatafileManager.returns({
+        datafileManager.HttpPollingDatafileManager.returns({
           start: sinon.stub(),
           stop: sinon.stub(),
           get: sinon.stub().returns(JSON.stringify(cloneDeep(configWithFeatures))),
@@ -229,7 +230,7 @@ describe('lib/core/project_config/project_config_manager', function() {
             variations: [{ key: 'variation', id: '99977477477747747' }],
           });
           nextDatafile.revision = '36';
-          var fakeDatafileManager = datafileManager.createHttpPollingDatafileManager.getCall(0).returnValue;
+          var fakeDatafileManager = datafileManager.HttpPollingDatafileManager.getCall(0).returnValue;
           fakeDatafileManager.get.returns(cloneDeep(nextDatafile));
           var updateListener = fakeDatafileManager.on.getCall(0).args[1];
           updateListener({ datafile: nextDatafile });
@@ -238,7 +239,7 @@ describe('lib/core/project_config/project_config_manager', function() {
       });
 
       it('calls onUpdate listeners after becoming ready, and after the datafile manager emits updates', function() {
-        datafileManager.createHttpPollingDatafileManager.returns({
+        datafileManager.HttpPollingDatafileManager.returns({
           start: sinon.stub(),
           stop: sinon.stub(),
           get: sinon.stub().returns(JSON.stringify(testData.getTestProjectConfigWithFeatures())),
@@ -253,7 +254,7 @@ describe('lib/core/project_config/project_config_manager', function() {
         return manager.onReady().then(function() {
           sinon.assert.calledOnce(onUpdateSpy);
 
-          var fakeDatafileManager = datafileManager.createHttpPollingDatafileManager.getCall(0).returnValue;
+          var fakeDatafileManager = datafileManager.HttpPollingDatafileManager.getCall(0).returnValue;
           var updateListener = fakeDatafileManager.on.getCall(0).args[1];
           var newDatafile = testData.getTestProjectConfigWithFeatures();
           newDatafile.revision = '36';
@@ -265,7 +266,7 @@ describe('lib/core/project_config/project_config_manager', function() {
       });
 
       it('can remove onUpdate listeners using the function returned from onUpdate', function() {
-        datafileManager.createHttpPollingDatafileManager.returns({
+        datafileManager.HttpPollingDatafileManager.returns({
           start: sinon.stub(),
           stop: sinon.stub(),
           get: sinon.stub().returns(JSON.stringify(testData.getTestProjectConfigWithFeatures())),
@@ -279,7 +280,7 @@ describe('lib/core/project_config/project_config_manager', function() {
           var onUpdateSpy = sinon.spy();
           var unsubscribe = manager.onUpdate(onUpdateSpy);
 
-          var fakeDatafileManager = datafileManager.createHttpPollingDatafileManager.getCall(0).returnValue;
+          var fakeDatafileManager = datafileManager.HttpPollingDatafileManager.getCall(0).returnValue;
           var updateListener = fakeDatafileManager.on.getCall(0).args[1];
           var newDatafile = testData.getTestProjectConfigWithFeatures();
           newDatafile.revision = '36';
@@ -303,7 +304,7 @@ describe('lib/core/project_config/project_config_manager', function() {
       it('fulfills its ready promise with an unsuccessful result when the datafile manager emits an invalid datafile', function() {
         var invalidDatafile = testData.getTestProjectConfig();
         delete invalidDatafile['projectId'];
-        datafileManager.createHttpPollingDatafileManager.returns({
+        datafileManager.HttpPollingDatafileManager.returns({
           start: sinon.stub(),
           stop: sinon.stub(),
           get: sinon.stub().returns(JSON.stringify(invalidDatafile)),
@@ -322,7 +323,7 @@ describe('lib/core/project_config/project_config_manager', function() {
       });
 
       it('fullfils its ready promise with an unsuccessful result when the datafile manager onReady promise rejects', function() {
-        datafileManager.createHttpPollingDatafileManager.returns({
+        datafileManager.HttpPollingDatafileManager.returns({
           start: sinon.stub(),
           stop: sinon.stub(),
           get: sinon.stub().returns(null),
@@ -345,7 +346,7 @@ describe('lib/core/project_config/project_config_manager', function() {
           sdkKey: '12345',
         });
         manager.stop();
-        sinon.assert.calledOnce(datafileManager.createHttpPollingDatafileManager.getCall(0).returnValue.stop);
+        sinon.assert.calledOnce(datafileManager.HttpPollingDatafileManager.getCall(0).returnValue.stop);
       });
 
       it('does not log an error message', function() {
@@ -358,7 +359,7 @@ describe('lib/core/project_config/project_config_manager', function() {
 
     describe('when constructed with sdkKey and with a valid datafile object', function() {
       it('fulfills its onReady promise with a successful result, and does not call onUpdate listeners after becoming ready', function() {
-        datafileManager.createHttpPollingDatafileManager.returns({
+        datafileManager.HttpPollingDatafileManager.returns({
           start: sinon.stub(),
           stop: sinon.stub(),
           get: sinon.stub().returns(JSON.stringify(testData.getTestProjectConfigWithFeatures())),
@@ -385,7 +386,7 @@ describe('lib/core/project_config/project_config_manager', function() {
 
     describe('when constructed with sdkKey and with a valid datafile string', function() {
       it('fulfills its onReady promise with a successful result, and does not call onUpdate listeners after becoming ready', function() {
-        datafileManager.createHttpPollingDatafileManager.returns({
+        datafileManager.HttpPollingDatafileManager.returns({
           start: sinon.stub(),
           stop: sinon.stub(),
           get: sinon.stub().returns(JSON.stringify(testData.getTestProjectConfigWithFeatures())),
@@ -430,7 +431,7 @@ describe('lib/core/project_config/project_config_manager', function() {
         manager.getOptimizelyConfig();
         sinon.assert.calledOnce(optimizelyConfig.createOptimizelyConfig);
         // create config with new revision
-        var fakeDatafileManager = datafileManager.createHttpPollingDatafileManager.getCall(0).returnValue;
+        var fakeDatafileManager = datafileManager.HttpPollingDatafileManager.getCall(0).returnValue;
         var updateListener = fakeDatafileManager.on.getCall(0).args[1];
         var newDatafile = testData.getTestProjectConfigWithFeatures();
         newDatafile.revision = '36';
